requirements.txt
# Core ML & Data Processing
pandas==2.1.3
polars==0.19.14
numpy==1.24.3
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.1
catboost==1.2.2

# Feature Engineering
featuretools==1.26.0

# Hyperparameter Optimization
optuna==3.4.0

# ML Tracking & Deployment
mlflow==2.8.1
wandb==0.16.0

# Database & Storage
duckdb==0.9.2
delta-lake==0.15.0
boto3==1.29.57

# Async HTTP & Scraping  
aiohttp==3.9.1
httpx==0.25.2
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# Orchestration
apache-airflow==2.7.3
prefect==2.14.11

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# GUI Framework
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2

# LLM Integration
openai==1.3.7
langchain==0.0.340
tiktoken==0.5.2

# Data Validation
great-expectations==0.18.5

# Security & Configuration
python-dotenv==1.0.0
cryptography==41.0.8

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Automation & Selenium (for bet execution)
selenium==4.15.2
webdriver-manager==4.0.1

# Timezone handling
pytz==2023.3

# Performance monitoring
psutil==5.9.6